refactor: Phase 2 — extract StandingsUpdaterRepository from StandingsRepository
